requirements.txt
ovh==1.0.0
python-dotenv==0.21.0
requests==2.22.0
